backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
duckdb>=0.9.2
pandas>=2.1.0
openpyxl>=3.1.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0.1
python-multipart>=0.0.6
cryptography>=41.0.0
python-dotenv>=1.0.0
apscheduler>=3.10.0

